+++ R/garnishMap.R
@@ -3,9 +3,9 @@
 #' @description
 #' This function provides a versatile interface to add components to a
 #' leaflet or mapview map. It takes functions such as "addMouseCoordinates"
-#' or \code{\link{addLayersControl}} and their repsective arguments and adds
+#' or \code{\link{addLayersControl}} and their respective arguments and adds
 #' them to the map. Arguments must be named. Functions can be plain or
-#' charatcer strings.
+#' character strings.
 #'
 #' @param map a mapview or leaflet object.
 #' @param ... functions and their arguments to add things to a map.
@@ -23,7 +23,7 @@
 #'
 #' ## even more flexible
 #' m2 <- garnishMap(m1, addPolygons, data = franconia, popup = popupTable(franconia),
-#'            fillOpacity = 0.8, color = "black", fillColor = "#BEBEBE")
+#'                  fillOpacity = 0.8, color = "black", fillColor = "#BEBEBE")
 #' garnishMap(m2, addCircleMarkers, data = breweries)
 #'
 #' }