+++ R/mapview-package.R
@@ -12,7 +12,7 @@
 #' Stephan Woellauer, Spaska Forteva, Thomas Nauss,
 #' Environmental Informatics Marburg \cr
 #' \cr
-#' \emph{Maintainer:} Tim Appelhans \email{admin@@environmentalinformatics-marburg.de}
+#' \emph{Maintainer:} Tim Appelhans \email{tim.appelhans@gmail.com}
 #'
 #' @import leaflet sp sf raster satellite scales Rcpp methods png lattice gdalUtils viridisLite base64enc
 #' @importFrom grDevices rgb dev.off png svg colorRampPalette